requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
recipe-scrapers>=15.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask-login>=0.6.0
